fix when reshuffling happens
the official rules (secrethitler.com) state:

	Top card case:
If there are fewer than three tiles remaining in the Policy deck at this point,
shuffle them with the Discard pile to create a new Policy deck.

	After a session:
If there are fewer than three tiles remaining in the Policy deck at the end
of a Legislative Session, shuffle them with the Discard pile to create
a new Policy deck.