fix bug where turn order would be wrong if a special election killed someone
the current president field was an index. After an execute the index would
be updated. But if the execute was during a special election then the turn
order would start after the specially elected president.

Fix by having the current president be the player's name and the next
president is updated at the end of the round.